@@ -12,24 +12,24 @@
         }
     ],
     "require": {
-        "php": ">=7.2",
-        "composer/composer": "~1.9.3",
-        "doctrine/annotations": "^1.8.0",
-        "doctrine/dbal": "^2.10.1",
+        "php": ">=7.3",
+        "composer/composer": "~1.10.5",
+        "doctrine/annotations": "^1.10.2",
+        "doctrine/dbal": "^2.10.2",
         "paragonie/random-lib": "^2.0.1",
-        "maximebf/debugbar": "^1.16",
+        "maximebf/debugbar": "^1.16.2",
         "monolog/monolog": "^2.0.2",
-        "nikic/php-parser": "~4.3",
-        "psr/log": "~1.1",
+        "nikic/php-parser": "~4.4.0",
+        "psr/log": "~1.1.3",
         "swiftmailer/swiftmailer": "~6.2",
-        "symfony/debug": "~4.4",
-        "symfony/finder": "~4.4",
-        "symfony/http-foundation": "~4.4",
-        "symfony/routing": "~4.4",
-        "symfony/stopwatch": "~4.4",
-        "symfony/templating": "~4.4",
-        "symfony/translation": "~4.4",
-        "twig/twig": "^2.12"
+        "symfony/debug": "~4.4.7",
+        "symfony/finder": "~5.0.7",
+        "symfony/http-foundation": "~5.0.7",
+        "symfony/routing": "~5.0.7",
+        "symfony/stopwatch": "~5.0.7",
+        "symfony/templating": "~5.0.7",
+        "symfony/translation": "~5.0.7",
+        "twig/twig": "~3.0.3"
     },
     "require-dev": {
         "ext-dom": "*",
@@ -38,7 +38,7 @@
         "ext-mbstring": "*",
         "ext-xml": "*",
         "ext-xmlwriter": "*",
-        "phpunit/phpunit": "~8.5"
+        "phpunit/phpunit": "~9.1.2"
     },
     "autoload": {
         "psr-4": {
