@@ -13,14 +13,14 @@ return [
 
         $app['twig'] = function ($app) {
 
-            $twig = new Twig_Environment(new TwigLoader(isset($app['locator']) ? new FilesystemLoader($app['locator']) : new SymfonyFilesystemLoader([])), [
+            $twig = new \Twig\Environment(new TwigLoader(isset($app['locator']) ? new FilesystemLoader($app['locator']) : new SymfonyFilesystemLoader([])), [
                 'cache' => new TwigCache($app['path.cache']),
                 'auto_reload' => true,
                 'debug' => $app['debug'],
             ]);
 
             if (isset($app['debug']) && $app['debug']) {
-                $twig->addExtension(new Twig_Extension_Debug());
+                $twig->addExtension(new \Twig\Extension\DebugExtension());
             }
 
             return $twig;
