@@ -161,7 +161,7 @@
 
 <script>
 
-import { ValidationObserver, VInput } from 'SystemApp/components/validation.vue';
+import { ValidationObserver, VInput } from '@system/app/components/validation.vue';
 
 var Installer = {
 
